Extract the Solution gallery into its own module
